faster device switching
